He eliminado las copias de la refactorización. Por lo que la parte de datos está modificada y solo faltaría crear sus tests. (Osea, obviamente no es perfecta. Pero ya la entiendo)
