--- src/configuration/__init__.py
@@ -1,35 +1,0 @@
-"""Módulo de configuración para el bot de trading."""
-
-from .config_model import (
-    AppConfig, 
-    EnvironmentConfig, 
-    AgentConfig, 
-    TrendIndicatorsConfig,
-    MomentumIndicatorsConfig,
-    VolatilityIndicatorsConfig,
-    VolumeIndicatorsConfig,
-    NormalizationConfig,
-    GCPConfig,
-    APIConfig,
-    TradingConfig,
-    LiveTradingConfig,
-    SystemConfig,
-    EvaluationConfig
-)
-
-__all__ = [
-    'AppConfig',
-    'EnvironmentConfig', 
-    'AgentConfig',
-    'TrendIndicatorsConfig',
-    'MomentumIndicatorsConfig', 
-    'VolatilityIndicatorsConfig',
-    'VolumeIndicatorsConfig',
-    'NormalizationConfig',
-    'GCPConfig',
-    'APIConfig',
-    'TradingConfig',
-    'LiveTradingConfig',
-    'SystemConfig',
-    'EvaluationConfig'
-]
