--- src/configuration/config.yaml
@@ -1,207 +1,0 @@
-# ----------------------------------------------------
-# SECCIÓN 1: CONFIGURACIÓN DEL ENTRENAMIENTO
-# ----------------------------------------------------
-training_setup:
-  seed: 73  # Semilla para la reproducibilidad del entrenamiento
-
-# ----------------------------------------------------
-# SECCIÓN 2: CONFIGURACIÓN DEL SISTEMA
-# ----------------------------------------------------
-# Configuración para el bot de trading de Bitcoin
-# Parámetros de la API de Binance
-api:
-  call_limit: 1000  # Número máximo de velas por llamada
-  max_retries: 3    # Número máximo de reintentos en caso de error
-  retry_delay: 1    # Segundos entre reintentos
-  timeout: 10       # Timeout para las peticiones en segundos
-
-# Configuración de trading
-trading:
-  testnet: true  # Si está en modo testnet (true) o producción (false)
-
-# Configuración de Google Cloud
-gcp:
-  project_id: "btcbot-2762"  # Tu project ID de Google Cloud
-  storage:
-    bucket_name: "btcbot-training-2762"  # Nombre del bucket para almacenar modelos
-    scaler_blob_name: "scaler.pkl"  # Nombre del archivo scaler en GCS
-    price_scaler_blob_name: "price_scaler.pkl"  # Nombre del archivo price_scaler en GCS
-  secrets:
-    # Nombres de los secretos tal como existen en Google Secret Manager
-    binance_api_key_futures: "BINANCE_API_KEY_FUTURES"
-    binance_api_secret_futures: "BINANCE_API_SECRET_FUTURES"
-    testnet_binance_api_key_futures: "TESTNET_BINANCE_API_KEY_FUTURES"
-    testnet_binance_api_secret_futures: "TESTNET_BINANCE_API_SECRET_FUTURES"
-    telegram_bot_token: "TELEGRAM_BOT_TOKEN"
-    telegram_chat_id: "TELEGRAM_CHAT_ID" 
-
-# Configuración de datos
-data:
-  # Columnas OHLCV a mantener
-  columns:
-    - "Open"
-    - "High" 
-    - "Low"
-    - "Close"
-    - "Volume"
-  
-  # Configuración de tipos de datos para optimizar RAM
-  dtypes:
-    Open: "float32"
-    High: "float32"
-    Low: "float32" 
-    Close: "float32"
-    Volume: "float32"
-
-# Configuración de zona horaria
-timezone:
-  target: "Europe/Madrid"  # UTC+1/UTC+2
-  source: "UTC"
-
-# Configuración de interpolación
-interpolation:
-  method: "linear"  # Método de interpolación para NaNs
-  limit_direction: "both"  # Dirección para fill
-
-# Configuración de normalización
-normalization:
-  scaler_type: "MinMaxScaler"  # Tipo de escalador a usar
-  feature_range: [0, 1]        # Rango de normalización
-  storage_mode: "local"        # Modo de almacenamiento: "local" (simplificado)
-
-# Indicadores de Tendencia
-trend_indicators:
-  ema_20:
-    period: 20
-    enabled: true
-  ema_50:
-    period: 50
-    enabled: true
-  adx:
-    period: 14
-    enabled: true
-
-# Indicadores de Momento
-momentum_indicators:
-  rsi:
-    period: 14
-    enabled: true
-  stoch:
-    k_period: 14
-    d_period: 3
-    smooth_k: 3
-    enabled: true
-
-# Indicadores de Volatilidad
-volatility_indicators:
-  atr:
-    period: 14
-    enabled: true
-
-# Indicadores de Volumen
-volume_indicators:
-  obv:
-    enabled: true
-
-# Configuración del entorno de trading (Environment)
-environment:
-  # Capital y riesgo
-  capital_inicial: 10000.0  # Capital inicial en USDT
-  apalancamiento: 10        # Apalancamiento máximo (ej. 10x)
-  porcentaje_max_inversion_por_trade: 0.05  # 5% del balance como margen por trade
-  max_drawdown_configurado_cuenta: 0.5      # 50% de drawdown máximo antes de terminated
-  max_consecutive_losses: 5  # Número máximo de pérdidas consecutivas permitidas
-  stop_loss_pct: 0.02       # Stop-loss al 2% de pérdida sobre el precio de entrada
-  
-  # Costos de trading
-  comision_taker_porcentaje: 0.0004    # 0.04% sobre valor nocional
-  slippage_porcentaje: 0.00005         # 0.005% sobre precio de ejecución
-  
-  # Observación
-  ventana_observacion_size: 24         # Número de velas pasadas en la observación (L)
-  
-  # Normalización de portfolio para observación
-  max_pasos_en_posicion: 100          # Máximo esperado de pasos en una posición
-  min_clip_pnl_roe: -1.0             # ROE mínimo para clip (-100%)
-  max_clip_pnl_roe: 1.0              # ROE máximo para clip (+100%)
-
-  # Zona muerta para acciones
-  zona_muerta_mantener: 0.05          # ±0.05 para mantener posición
-
-  # Función de recompensa
-  peso_recompensa_paso: 1.0           # Peso para recompensa por paso
-  peso_recompensa_cierre: 1.0         # Peso para recompensa al cerrar operación
-  peso_recompensa_episodio: 0.0       # Peso para Sortino (0 = no se usa por ahora)
-  reward_strategy: "Sortino"          # Estrategia de recompensa: 'EquityChange' o 'Sortino'
-
-  # Condiciones de finalización
-  usar_max_pasos_episodio: false      # Si limitar pasos por episodio
-  max_pasos_episodio: 2000           # Máximo pasos por episodio (si se usa)
-
-# Configuración del Agente SAC con Transformer
-agent:
-  algorithm: "SAC"
-
-  # --- Arquitectura de la Red ---
-  architecture:
-    portfolio_features: 4
-    transformer_max_seq_len: 100
-    positional_encoding_learnable: false
-
-  # --- Parámetros del Replay Buffer y Entrenamiento General ---
-  replay_buffer_size: 1000000
-  batch_size: 8192
-  min_buffer_for_learning: 1
-
-  # --- Hiperparámetros Específicos de SAC ---
-  hiperparametros_sac:
-    gamma: 0.99
-    tau: 0.005
-    actor_learning_rate: 0.0003
-    critic_learning_rate: 0.0003
-    learn_alpha: True
-    alpha_learning_rate: 0.0003
-    target_entropy: 'auto'
-    initial_log_alpha: 0.0
-    log_std_min: -20  # Límite inferior para log_std
-    log_std_max: 2    # Límite superior para log_std
-    # Prioritized Experience Replay parameters
-    per_alpha: 0.6    # Priority exponent (0 = uniform random, 1 = fully prioritized)
-    per_beta: 0.4     # Importance sampling exponent (0 = no correction, 1 = full correction)
-
-  # --- Parámetros de la Arquitectura del Transformer ---
-  transformer:
-    d_model: 128
-    n_head: 4
-    num_encoder_layers: 3
-    dim_feedforward: 256
-    dropout_rate: 0.1
-
-  # --- Parámetros de las Cabezas MLP ---
-  mlp_heads:
-    hidden_dims: [256, 256]
-
-  # --- Frecuencia de Entrenamiento ---
-  learning_frequency: 1
-  update_target_frequency: 1
-
-  # --- Parámetros de Guardado y Carga ---
-  model_save_frequency: 10000
-  models_directory: "models/agent"
-
-# ----------------------------------------------------
-# SECCIÓN 3: CONFIGURACIÓN DE EJECUCIÓN
-# ----------------------------------------------------
-live_trading:
-  warm_up_candles: 300
-  default_model_to_load: "best_model"
-  bigquery_dataset_id: "trading_logs"
-  bigquery_table_id: "live_trading_log"
-
-system:
-  max_parallel_workers_api: 8
-  nccl_socket_ifname: 'eth0'
-
-evaluation:
-  temp_directory: "temp_evaluation"
-  default_model_to_load: "best"
