+++ src/data/__init__.py
@@ -3,9 +3,11 @@
 from .abstractions import DataSource
 from .binance_source import BinanceDataSource
 from .pipeline import DataPipeline
+from .artifact_manager import ArtifactManager
 
 __all__ = [
     'DataSource',
     'BinanceDataSource', 
-    'DataPipeline'
+    'DataPipeline',
+    'ArtifactManager'
 ]
