+++ src/data/pipeline.py
@@ -4,22 +4,22 @@ Unifica todo el preprocesamiento de datos incluyendo adquisición, indicadores y
 """
 
 import logging
-from typing import Tuple, Dict, Optional
+from typing import Tuple
 import pandas as pd
-from pathlib import Path
 
-from .binance_source import BinanceDataSource
 from .abstractions import DataSource
 from .indicadores import Indicadores
 from .normalization import Normalization
 
+from src.config import (
+    AppConfig
+)
 
 class DataPipeline:
     """Clase que unifica todo el preprocesamiento de datos."""
     
-    def __init__(self, data_source: DataSource, symbol: str, interval: str, start_date: str, run_id: str, base_path: str,
-                 full_config: Dict, end_date: str = None, save_artifacts: bool = True, 
-                 gcs_utils=None):
+    def __init__(self, data_source: DataSource, symbol: str, interval: str, start_date: str, end_date: str, run_id: str, base_path: str,
+                 config: AppConfig):
         """
         Inicializa el pipeline de datos.
         
@@ -32,8 +32,6 @@ class DataPipeline:
             base_path (str): Ruta base para guardar los artifacts del entrenamiento
             full_config (Dict): Configuración completa para todas las clases del pipeline
             end_date (str, optional): Fecha de fin en formato YYYY-MM-DD
-            save_artifacts (bool): Si True, guarda artefactos como scalers
-            gcs_utils: Instancia de GCSUtils para operaciones en la nube (opcional)
         """
         # Inyección de dependencia: fuente de datos
         self.data_source = data_source
@@ -44,23 +42,19 @@ class DataPipeline:
         self.end_date = end_date
         self.run_id = run_id
         self.base_path = base_path
-        self.save_artifacts = save_artifacts
         
         # Store injected configuration and dependencies
-        self.full_config = full_config
-        self.gcs_utils = gcs_utils
+        self.config = config
         
         # Configurar logging
-        logging.basicConfig(level=logging.INFO)
         self.logger = logging.getLogger(__name__)
         
         self.logger.info(f"Pipeline de datos inicializado para {symbol} ({interval})")
         self.logger.info(f"Período: desde {start_date}" + (f" hasta {end_date}" if end_date else " hasta ahora"))
         self.logger.info(f"Run ID: {run_id}")
         self.logger.info(f"Base path: {base_path}")
-        self.logger.info(f"Guardar artefactos: {save_artifacts}")
-        self.logger.info(f"Configuración inyectada: {len(self.full_config)} secciones")
-    
+        self.logger.info(f"Configuración inyectada: {len(self.config.model_dump())} secciones")
+
     def run(self) -> Tuple[pd.DataFrame, object, object]:
         """
         Ejecuta el pipeline completo de preprocesamiento de datos.
@@ -84,7 +78,7 @@ class DataPipeline:
         self.logger.info("PASO 2: Cálculo de Indicadores Técnicos")
         indicadores = Indicadores(
             dataframe, 
-            config_dict=self.full_config
+            config=self.config
         )
         dataframe_with_indicators = indicadores.main()
         
@@ -103,13 +97,11 @@ class DataPipeline:
         # Paso 3: Normalización de datos
         self.logger.info("PASO 3: Normalización de Datos")
         normalization = Normalization(
-            dataframe_with_indicators, 
-            base_path=self.base_path, 
-            run_id=self.run_id,
-            save_artifacts=False,  # ArtifactManager se encargará del guardado
-            normalization_config=self.full_config.get('normalization', {}),
-            gcs_utils=self.gcs_utils
-        )
+            dataframe_with_indicators,
+            config=self.config,
+            base_path=self.base_path,
+            run_id=self.run_id
+            )
         normalized_dataframe, scaler, price_scaler = normalization.main()
         
         self.logger.info(f"Normalización completada exitosamente:")
@@ -120,12 +112,6 @@ class DataPipeline:
         self.logger.info("  - Artefactos disponibles para guardado por ArtifactManager")
             
         self.logger.info(f"  - Memoria utilizada: {normalized_dataframe.memory_usage(deep=True).sum() / 1024**2:.2f} MB")
-        
-        # Mostrar información del scaler
-        feature_info = normalization.get_feature_info()
-        self.logger.info(f"  - Características normalizadas: {feature_info['num_features']}")
-        self.logger.info(f"  - Tipo de scaler: {feature_info['scaler_type']}")
-        self.logger.info(f"  - Rango de normalización: {feature_info['feature_range']}")
         
         self.logger.info("=== Pipeline de Datos Completado ===")
